Collect collections into one component
