+++ components/lark-collections/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2018"
 [dependencies]
 rustc-hash = "1.0"
 indexmap = "1.0.2"
+lark-debug-with = { path = "../lark-debug-with" }
