+++ components/lark-entity/Cargo.toml
@@ -5,10 +5,10 @@ authors = ["Niko Matsakis <niko@alum.mit.edu>"]
 edition = "2018"
 
 [dependencies]
+lark-collections = { path = "../lark-collections" }
 lark-debug-derive = { path = "../lark-debug-derive" }
 lark-debug-with = { path = "../lark-debug-with" }
 lark-error = { path = "../lark-error" }
-lark-indices = { path = "../lark-indices" }
 lark-intern = { path = "../lark-intern" }
 lark-string = { path = "../lark-string" }
 lark-span = { path = "../lark-span" }
