+++ components/lark-intern/Cargo.toml
@@ -6,5 +6,4 @@ edition = "2018"
 
 [dependencies]
 lark-collections = { path = "../lark-collections" }
-lark-indices = { path = "../lark-indices" }
 parking_lot = "0.7"
