+++ components/lark-intern/src/lib.rs
@@ -1,5 +1,4 @@
-use lark_collections::{Equivalent, FxIndexMap};
-use lark_indices::U32Index;
+use lark_collections::{Equivalent, FxIndexMap, U32Index};
 use parking_lot::{RwLock, RwLockUpgradableReadGuard};
 use std::hash::Hash;
