+++ components/lark-parser/src/lib.rs
@@ -9,7 +9,7 @@
 use crate::lexer::token::LexToken;
 use crate::macros::EntityMacroDefinition;
 use crate::syntax::entity::ParsedEntity;
-use lark_collections::FxIndexMap;
+use lark_collections::{FxIndexMap, Seq};
 use lark_debug_derive::DebugWith;
 use lark_entity::Entity;
 use lark_entity::EntityData;
@@ -20,7 +20,6 @@ use lark_error::ErrorReported;
 use lark_error::WithError;
 use lark_hir as hir;
 use lark_intern::Intern;
-use lark_seq::Seq;
 use lark_span::ByteIndex;
 use lark_span::FileName;
 use lark_span::IntoFileName;
