+++ components/lark-parser/src/macros/function_declaration.rs
@@ -17,6 +17,7 @@ use crate::syntax::sigil::{Curlies, Parentheses, RightArrow};
 use crate::syntax::skip_newline::SkipNewline;
 use crate::syntax::type_reference::ParsedTypeReference;
 use crate::syntax::type_reference::TypeReference;
+use lark_collections::Seq;
 use lark_debug_with::DebugWith;
 use lark_entity::Entity;
 use lark_entity::EntityData;
@@ -28,7 +29,6 @@ use lark_error::WithError;
 use lark_hir as hir;
 use lark_intern::Intern;
 use lark_intern::Untern;
-use lark_seq::Seq;
 use lark_span::FileName;
 use lark_span::Spanned;
 use lark_string::GlobalIdentifier;
