+++ components/lark-parser/src/macros/struct_declaration.rs
@@ -10,6 +10,7 @@ use crate::syntax::identifier::SpannedGlobalIdentifier;
 use crate::syntax::list::CommaList;
 use crate::syntax::sigil::Curlies;
 use crate::syntax::skip_newline::SkipNewline;
+use lark_collections::Seq;
 use lark_debug_with::DebugWith;
 use lark_entity::Entity;
 use lark_entity::EntityData;
@@ -20,7 +21,6 @@ use lark_error::ErrorSentinel;
 use lark_error::WithError;
 use lark_hir as hir;
 use lark_intern::Intern;
-use lark_seq::Seq;
 use lark_span::FileName;
 use lark_span::Spanned;
 use lark_string::GlobalIdentifier;
