+++ components/lark-parser/src/parser.rs
@@ -3,11 +3,10 @@ use crate::macros::EntityMacroDefinition;
 use crate::syntax::NonEmptySyntax;
 use crate::syntax::Syntax;
 
-use lark_collections::FxIndexMap;
+use lark_collections::{FxIndexMap, Seq};
 use lark_debug_with::DebugWith;
 use lark_entity::EntityTables;
 use lark_error::{Diagnostic, ErrorReported, WithError};
-use lark_seq::Seq;
 use lark_span::{FileName, Span, Spanned};
 use lark_string::{GlobalIdentifier, GlobalIdentifierTables, Text};
 use std::sync::Arc;
