+++ components/lark-parser/src/syntax/entity.rs
@@ -4,6 +4,7 @@ use crate::syntax::identifier::SpannedGlobalIdentifier;
 use crate::syntax::NonEmptySyntax;
 use crate::syntax::Syntax;
 use crate::ParserDatabase;
+use lark_collections::Seq;
 use lark_debug_derive::DebugWith;
 use lark_debug_with::DebugWith;
 use lark_entity::Entity;
@@ -12,7 +13,6 @@ use lark_error::ErrorReported;
 use lark_error::ErrorSentinel;
 use lark_error::WithError;
 use lark_hir as hir;
-use lark_seq::Seq;
 use lark_span::FileName;
 use lark_span::Span;
 use lark_span::Spanned;
