+++ components/lark-parser/src/syntax/field.rs
@@ -9,13 +9,13 @@ use crate::syntax::sigil::Colon;
 use crate::syntax::skip_newline::SkipNewline;
 use crate::syntax::type_reference::{ParsedTypeReference, TypeReference};
 use crate::syntax::Syntax;
+use lark_collections::Seq;
 use lark_debug_derive::DebugWith;
 use lark_entity::Entity;
 use lark_error::ErrorReported;
 use lark_error::ResultExt;
 use lark_error::WithError;
 use lark_hir as hir;
-use lark_seq::Seq;
 use lark_span::FileName;
 use lark_span::Spanned;
 use lark_string::GlobalIdentifier;
