+++ components/lark-parser/src/syntax/fn_body.rs
@@ -20,7 +20,7 @@ use crate::syntax::sigil::Semicolon;
 use crate::syntax::skip_newline::SkipNewline;
 use crate::syntax::Syntax;
 use derive_new::new;
-use lark_collections::FxIndexMap;
+use lark_collections::{FxIndexMap, Seq};
 use lark_debug_derive::DebugWith;
 use lark_debug_with::DebugWith;
 use lark_entity::Entity;
@@ -29,7 +29,6 @@ use lark_error::WithError;
 use lark_hir as hir;
 use lark_intern::Intern;
 use lark_intern::Untern;
-use lark_seq::Seq;
 use lark_span::FileName;
 use lark_span::Span;
 use lark_span::Spanned;
