--- components/lark-seq/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "lark-seq"
-version = "0.1.0"
-authors = ["Niko Matsakis <niko@alum.mit.edu>"]
-edition = "2018"
-
-[dependencies]
-lark-debug-with = { path = "../lark-debug-with" }
