+++ components/lark-test/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 diff = "0.1.11"
 env_logger = "0.6"
 languageserver-types = "0.53"
+lark-collections = { path = "../lark-collections" }
 lark-debug-with = { path = "../lark-debug-with" }
 lark-cli = { path = "../lark-cli" }
 lark-entity = { path = "../lark-entity" }
@@ -18,12 +19,10 @@ lark-intern = { path = "../lark-intern" }
 lark-language-server = { path = "../lark-language-server" }
 lark-query-system = { path = "../lark-query-system" }
 lark-parser = { path = "../lark-parser" }
-lark-seq = { path = "../lark-seq" }
 lark-span = { path = "../lark-span" }
 lark-string = { path = "../lark-string" }
 lark-ty = { path = "../lark-ty" }
 lark-type-check = { path = "../lark-type-check" }
-lark-collections = { path = "../lark-collections" }
 lazy_static = "1.2.0"
 rayon = "1.0.3"
 regex = "1"
