+++ components/lark-test/src/lib.rs
@@ -6,7 +6,7 @@ use lark_intern::Intern;
 use lark_parser::{ParserDatabase, ParserDatabaseExt};
 use lark_query_system::ls_ops::{Cancelled, LsDatabase, RangedDiagnostic};
 use lark_query_system::LarkDatabase;
-use lark_seq::seq;
+use lark_collections::seq;
 use lark_span::FileName;
 use lark_string::Text;
 use salsa::Database;
