+++ components/lark-ty/Cargo.toml
@@ -11,10 +11,8 @@ parking_lot = "0.7"
 lark-collections = { path = "../lark-collections" }
 lark-debug-derive = { path = "../lark-debug-derive" }
 lark-debug-with = { path = "../lark-debug-with" }
-lark-indices = { path = "../lark-indices" }
 lark-intern = { path = "../lark-intern" }
 lark-error = { path = "../lark-error" }
 lark-entity = { path = "../lark-entity" }
-lark-seq = { path = "../lark-seq" }
 lark-string = { path = "../lark-string" }
 lark-unify = { path = "../lark-unify" }
