+++ components/lark-ty/src/base_inferred.rs
@@ -38,7 +38,7 @@ impl TypeFamily for BaseInferred {
     }
 }
 
-lark_indices::index_type! {
+lark_collections::index_type! {
     pub struct Base { .. }
 }
