+++ components/lark-ty/src/full_inferred.rs
@@ -39,7 +39,7 @@ impl TypeFamily for FullInferred {
     }
 }
 
-lark_indices::index_type! {
+lark_collections::index_type! {
     pub struct Base { .. }
 }
