+++ components/lark-ty/src/map_family.rs
@@ -7,10 +7,8 @@ use crate::Generics;
 use crate::Signature;
 use crate::Ty;
 use crate::TypeFamily;
-use lark_collections::FxIndexMap;
-use lark_collections::FxIndexSet;
+use lark_collections::{FxIndexMap, FxIndexSet, Seq};
 use lark_entity::Entity;
-use lark_seq::Seq;
 use std::collections::BTreeMap;
 use std::hash::Hash;
 use std::sync::Arc;
