+++ components/lark-type-check/Cargo.toml
@@ -12,7 +12,6 @@ generational-arena = "0.2.0"
 lark-collections = { path = "../lark-collections" }
 lark-debug-derive = { path = "../lark-debug-derive" }
 lark-debug-with = { path = "../lark-debug-with" }
-lark-indices = { path = "../lark-indices" }
 lark-intern = { path = "../lark-intern" }
 lark-error = { path = "../lark-error" }
 lark-entity = { path = "../lark-entity" }
