+++ components/lark-type-check/src/base_inference/query_definition.rs
@@ -5,11 +5,10 @@ use crate::TypeCheckDatabase;
 use crate::TypeChecker;
 use crate::UniverseBinder;
 use generational_arena::Arena;
-use lark_collections::FxIndexMap;
+use lark_collections::{FxIndexMap, IndexVec};
 use lark_entity::Entity;
 use lark_error::Diagnostic;
 use lark_error::WithError;
-use lark_indices::IndexVec;
 use lark_ty::base_inferred::BaseInferred;
 use lark_ty::map_family::Map;
 use lark_unify::UnificationTable;
