+++ components/lark-type-check/src/full_inference/query_definition.rs
@@ -8,11 +8,10 @@ use crate::TypeCheckDatabase;
 use crate::TypeChecker;
 use crate::UniverseBinder;
 use generational_arena::Arena;
-use lark_collections::FxIndexMap;
+use lark_collections::{FxIndexMap, IndexVec};
 use lark_entity::Entity;
 use lark_error::Diagnostic;
 use lark_error::WithError;
-use lark_indices::IndexVec;
 use lark_ty::full_inferred::FullInferred;
 use lark_ty::map_family::Map;
 use lark_unify::UnificationTable;
