+++ components/lark-type-check/src/full_inference/type_checker.rs
@@ -13,10 +13,9 @@ use crate::TypeCheckDatabase;
 use crate::TypeChecker;
 use crate::TypeCheckerFamilyDependentExt;
 use crate::TypeCheckerVariableExt;
-use lark_collections::FxIndexSet;
+use lark_collections::{FxIndexSet, IndexVec};
 use lark_entity::Entity;
 use lark_hir as hir;
-use lark_indices::IndexVec;
 use lark_intern::Intern;
 use lark_intern::Untern;
 use lark_pretty_print::PrettyPrint;
