+++ components/lark-unify/Cargo.toml
@@ -5,5 +5,5 @@ authors = ["Niko Matsakis <niko@alum.mit.edu>"]
 edition = "2018"
 
 [dependencies]
+lark-collections = { path = "../lark-collections" }
 lark-debug-with = { path = "../lark-debug-with" }
-lark-indices = { path = "../lark-indices" }
